--- editoast/fga/src/client/healthz.rs
@@ -1,1 +1,0 @@
-pub(super) use super::api::healthz::Health;
